improved precision for geometric mean return

+++ calc/src/test/java/org/javamoney/calc/securities/GeometricMeanReturnTest.java
@@ -2,6 +2,7 @@ package org.javamoney.calc.securities;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -12,7 +13,7 @@ import static org.junit.Assert.assertEquals;
  */
 public class GeometricMeanReturnTest {
 
-    private static final List<Double> RATES_OF_RETURN = Arrays.asList(0.2, 0.06, 0.01);
+    private static final List<BigDecimal> RATES_OF_RETURN = Arrays.asList(BigDecimal.valueOf(0.2), BigDecimal.valueOf(0.06), BigDecimal.valueOf(0.01));
     private static final int NUMBER_OF_PERIODS = 3;
 
     @Test
@@ -22,6 +23,6 @@ public class GeometricMeanReturnTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void testCalculateWithNullRatesThrowsException() {
-        GeometricMeanReturn.calculate(Arrays.asList(0.1, 0.1, null, 0.5), NUMBER_OF_PERIODS);
+        GeometricMeanReturn.calculate(Arrays.asList(BigDecimal.valueOf(0.1), BigDecimal.valueOf(0.1), null, BigDecimal.valueOf(0.5)), NUMBER_OF_PERIODS);
     }
 }
